refactor Comment component
