+++ package.json
@@ -14,6 +14,7 @@
     "babel-loader": "^6.2.8",
     "babel-preset-es2015": "^6.18.0",
     "babel-preset-react": "^6.16.0",
+    "jquery": "^3.1.1",
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
     "webpack": "^1.13.3",
