+++ public/comments.js
@@ -1,0 +1,6 @@
+[{"id": 1, "author": "Ben Rub", "body": "just wanna comment and hope it shows...", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" },
+{"id": 2, "author": "Bill Gates", "body": "and just another comment in no time", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" },
+{"id": 3, "author": "Ben Rub", "body": "just wanna comment and hope it shows...", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" },
+{"id": 4, "author": "Bill Gates", "body": "and just another comment in no time", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" },
+{"id": 5, "author": "Ben Rub", "body": "just wanna comment and hope it shows...", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" },
+{"id": 6, "author": "Bill Gates", "body": "and just another comment in no time", "avatarUrl": "http://www.avatar-zone.com/Avatars/Fun/Avatar-25.jpg" }]
